backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
websockets==12.0
python-socketio==5.11.1
python-engineio==4.9.0
aiohttp==3.9.3
faker==22.6.0
python-dateutil==2.8.2
pydantic-settings==2.2.1
httpx==0.27.0